logica paciente y corrección logica medico

+++ src/medico/medico.service.ts
@@ -39,6 +39,9 @@ export class MedicoService {
         if (!medico) {
             throw new BusinessLogicException('Médico no encontrado', BusinessError.NOT_FOUND);
         }
+        if (medico.pacientes && medico.pacientes.length > 0) {
+            throw new BusinessLogicException('No se puede eliminar un médico con pacientes asociados', BusinessError.BAD_REQUEST);
+        }
         await this.medicoRepository.remove(medico);
     }
 }
